nexcart_backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
setuptools==80.7.1
sqlparse==0.5.3
stripe==7.8.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
